Change `dependencies` to `devDependencies`
Later on it may be more advantageous to refactor
the project and add a server side to it; for now
I would like to ensure that it is smooth to do
so.

Also, Maestro is moving away from a static site
towards a more build-centric system, using Gulp
as well as Browserify and other tools. Having
these tools that are involved in the build
process present only when necessary should be
advantageous in terms of overall project
organization.